Add new tasks

+++ white/tasks/src/week02/task02/IsPalindrom.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 bool IsPalindrom(string s)
 {
-  for (size_t i = 0; i < s.size() / 2; ++i)
+  for (int i = 0; i < s.size() / 2; ++i)
   {
     if (s[i] != s[s.size() - i - 1])
     {
