+++ white/tasks/src/week02/task03/main.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 bool IsPalindrom(string s)
 {
-  for (size_t i = 0; i < s.size() / 2; ++i)
+  for (int i = 0; i < s.size() / 2; ++i)
   {
     if (s[i] != s[s.size() - i - 1])
     {
