+++ white/tasks/src/week02/task06/main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 void Reverse(std::vector<int> &source)
 {
-  for (size_t i = 0; i < source.size() / 2; i++)
+  for (int i = 0; i < source.size() / 2; i++)
   {
     int temp = source[source.size() - i - 1];
     source[source.size() - i - 1] = source[i];
