+++ white/tasks/src/week02/task06/reverse.cpp
@@ -2,7 +2,7 @@
 
 void Reverse(std::vector<int> &source)
 {
-  for (size_t i = 0; i < source.size() / 2; i++)
+  for (int i = 0; i < source.size() / 2; i++)
   {
     int temp = source[source.size() - i - 1];
     source[source.size() - i - 1] = source[i];
